nightly.conf: Update to ssh.gitlab.freedesktop.org
Add the old url to the old array, and the new one. This will prompt the
user with the following message when `dim updated-branches` is called:

	Update drm to new ssh://git@ssh.gitlab.freedesktop.org/drm/kernel.git? (y/N)

... for each repository it's changing. After that, git should be
configured with the new repositories.

Note that we are taking the opportunity to be explicit on the protocol,
adding the ssh:// and dropping the entry without it. It's clearer and
less entries to maintain.

Signed-off-by: Lucas De Marchi <lucas.demarchi@intel.com>
Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Tested-by: Jani Nikula <jani.nikula@intel.com>